+++ addon/components/flexberry-lookup.js
@@ -7,6 +7,8 @@ import { inject as service } from '@ember/service';
 import { computed, observer } from '@ember/object';
 import { later, run } from '@ember/runloop';
 import { merge } from '@ember/polyfills';
+import { isNone } from '@ember/utils';
+import { on } from '@ember/object/evented';
 import FlexberryBaseComponent from './flexberry-base-component';
 
 import { translationMacro as t } from 'ember-i18n';
@@ -98,6 +100,15 @@ export default FlexberryBaseComponent.extend({
   placeholder: t('components.flexberry-lookup.placeholder'),
 
   /**
+    Text on button preview value.
+
+    @property previewText
+    @type String
+    @default '<i class="eye icon"></i>'
+  */
+  previewText: '<i class="eye icon"></i>',
+
+  /**
     Text on button opening a modal window.
 
     @property chooseText
@@ -116,6 +127,14 @@ export default FlexberryBaseComponent.extend({
     @default '<i class="remove icon"></i>'
   */
   removeText: '<i class="remove icon"></i>',
+
+  /**
+    CSS classes for preview button.
+
+    @property previewButtonClass
+    @type String
+  */
+  previewButtonClass: undefined,
 
   /**
     CSS classes for choose button.
@@ -177,17 +196,64 @@ export default FlexberryBaseComponent.extend({
   */
   dropdown: false,
 
+  /* eslint-disable  ember/no-on-calls-in-components */
   /**
     Handles changes in dropdown.
 
     @method _dropdownObserver
     @private
   */
-  _dropdownObserver: observer('dropdown', function() {
+  _dropdownObserver: on('init', observer('dropdown', function() {
     if (this.get('dropdown')) {
       this._onDropdown();
     }
-  }),
+  })),
+  /* eslint-enable  ember/no-on-calls-in-components */
+
+  /**
+    Flag to show in lookup preview button.
+
+    @property showPreviewButton
+    @type Boolean
+    @default false
+  */
+  showPreviewButton: false,
+
+  /**
+    The URL that will be used for viewing the selected object.
+
+    @property previewFormRoute
+    @type String
+    @default undefined
+  */
+  previewFormRoute: undefined,
+
+  /**
+    Projection name preview form.
+
+    @property previewFormProjection
+    @type String
+    @default undefined
+  */
+  previewFormProjection: undefined,
+
+  /**
+    Flag to show the selected object in separate route.
+
+    @property previewOnSeparateRoute
+    @type Boolean
+    @default false
+  */
+  previewOnSeparateRoute: false,
+
+  /**
+    The controller for viewing the selected object.
+
+    @property controllerForPreview
+    @type String
+    @default undefined
+  */
+  controllerForPreview: undefined,
 
   /**
     Flag to show that lookup has search or autocomplete in dropdown mode.
@@ -608,6 +674,46 @@ export default FlexberryBaseComponent.extend({
       } else {
         this.get('currentController').send(remove, removeData);
       }
+    },
+
+    /**
+      Show window for select value.
+
+      @method actions.preview
+    */
+    preview() {
+      let previewFormRoute = this.get('previewFormRoute');
+      if (isNone(previewFormRoute)) {
+        throw new Error('`previewFormRoute` is undefined.');
+      }
+
+      let relatedModel = this.get('relatedModel');
+      let relationName = this.get('relationName');
+      let relationModelName = getRelationType(relatedModel, relationName);
+
+      let thisRouteName = this.get('currentController.routeName');
+      let thisRecordId = this.get('relatedModel.id');
+      let transitionOptions = {
+        queryParams: {
+          readonly: true,
+          parentRoute: thisRouteName,
+          parentRouteRecordId: thisRecordId
+        }
+      };
+
+      let previewData = {
+        recordId: this.get('value.id'),
+        transitionRoute: previewFormRoute,
+        transitionOptions: transitionOptions,
+        showInSeparateRoute: this.get('previewOnSeparateRoute'),
+        modelName: relationModelName,
+        controller: this.get('controllerForPreview'),
+        projection: this.get('previewFormProjection')
+      };
+
+      /* eslint-disable ember/closure-actions */
+      this.sendAction('preview', previewData);
+      /* eslint-enable ember/closure-actions */
     }
   },