+++ CHANGELOG.md
@@ -91,6 +91,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * `flexberry-dropdown` component:
     * Instead of many properties one `settings` property for manage `Semantic UI` settings.
 
+## [2.1.0-beta.0] - 2018-12-13
+### Added
+* `flexberry-lookup` component:
+    * Add preview button.
+* `flexberry-file` component:
+    * Added the ability to open files instead of downloading.
+
 ## [2.0.0] - 2018-12-07
 ### Added
 * `flexberry-objectlistview-route` mixin: